requirements.txt
flask>=2.3.0
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
nltk>=3.8.1
tqdm>=4.65.0
python-dotenv>=1.0.0
together>=0.2.5
huggingface-hub>=0.28.0
supabase>=2.0.0 